direwolf: add 'noautostart' privisioning config
This change adds `provisioning-noautostart.json`, which will create the
server with the minecraft service disabled. This is helpful for
restoring a backup, i.e. there's no need to create a new world on
start-up.

The `restore.sh` script was amended to enable the service after restore
is complete instead of just starting it.

Signed-off-by: Thilo Fromm <thilo.alexander@gmail.com>